requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Mail==0.10.0
WTForms==3.2.1
APScheduler==3.10.4
google-generativeai==0.8.3
python-dotenv==1.0.1
Werkzeug==3.1.3
email-validator==2.2.0
